% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listselector-package.R
\docType{package}
\name{listselector-package}
\alias{listselector-package}
\alias{listselector}
\title{List element selector}
\description{
Tools for selected elements from nested lists via a CSS selector like syntax.
}
\author{
Colin Rundel \email{rundel@gmail.com}
}
\keyword{package}
